skills/jordy33__binance_mcp__root/requirements.txt
python-binance>=1.0.19
web3>=6.11.1
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
mcp>=0.3.0
setuptools>=70.0.0